Minimism

+++ src/sprites/Food.java
@@ -5,7 +5,7 @@ import java.awt.Graphics2D;
 public class Food extends Sprite {
 	Boolean meat;
 
-	public Food(double x, double y, double size1, Boolean meat, int range) {
+	public Food(double x, double y, float size1, Boolean meat, int range) {
 		super(x, y, size1, meat ? 1000 : 10000, range, meat ? Color.PINK : Color.GREEN);
 		this.meat = meat;
 		data = meat;
